ai/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
numpy
psycopg2-binary
